Move match maker data packets into packet mod Fixes #111
