+++ Match Maker Server/src/data/mod.rs
@@ -1,23 +1,5 @@
-/// Data Packet, send and received by the server
-mod packet;
-pub use packet::*;
-
-/// Match Making Request (Incoming packet)
-mod match_making_request;
-pub use match_making_request::*;
-
-/// Match Making Response (Outgoing packet)
-mod match_making_response;
-pub use match_making_response::*;
-
-/// Match Making Queue (Internal queue)
-mod match_making_queue;
-pub use match_making_queue::*;
-
-/// Match Making Update (Outgoing packet)
-mod match_making_update;
-pub use match_making_update::*;
+/// Data Packets, send and received by the server
+pub mod packet;
 
 /// Internal structure for Peers
-mod peer;
-pub use peer::*;
+pub mod peer;
