+++ Match Maker Server/src/lib.rs
@@ -1,7 +1,5 @@
 /// Data related
-mod data;
-pub use data::*;
+pub mod data;
 
 /// App Configuration
-mod app_config;
-pub use app_config::*;
+pub mod app_config;
